changed namespace and moved parser & lexer
